feat: in-QGIS dialog for importing engine test events (CSV)
Adds an "Import engine test events (CSV)" toolbar action that opens a
dialog for bulk-loading engine-test event rows into the currently-open
ALAQS project. The dialog is a UI wrapper around the same core logic
the CLI script uses; users can now do this workflow without touching
the command line.

Motivation
----------

Phase 4a shipped only a CLI (scripts/import_engine_test_events.py)
because bulk-loading events felt like a scripting task. But the rest
of the engine-test workflow is UI-driven — users tick the "Engine
test site" checkbox in the QGIS area-source form, so it's confusing
to then have to open a terminal for the next step. This closes that
UX gap.

Refactor to share code between CLI and UI
-----------------------------------------

The QGIS-plugin ZIP only contains files under ``open_alaqs/`` (see
``.qgis-plugin-ci``: ``plugin_path: open_alaqs``). ``scripts/`` is
NOT packaged, so the dialog can't import from the CLI script at
runtime.

Moved the importer's core logic (types, read_csv, validate_csv_rows,
validate_against_db, apply_insert, format_summary, format_apply_summary)
to a plugin module at
``open_alaqs/core/tools/engine_test_import.py``. The CLI script is
now a thin argparse + stdout wrapper (~180 lines from ~500). Both
entry points import from the shared core module so they can never
diverge.

The CLI script re-exports the moved names so any existing external
code (including the 39-test CLI test file) works unchanged.

Files
-----

* open_alaqs/core/tools/engine_test_import.py (new, ~450 lines)
  - Full importer core: CSV format documentation, all data types
    (RowIssue, ValidatedRow, ValidationResult), all validation
    codes (8 errors, 5 warnings), all three insert modes
    (append / replace-for-source / replace-all).
  - Dependency-free (stdlib only). QGIS-free.
  - Public API is ``__all__``.

* scripts/import_engine_test_events.py (reduced from ~500 to ~180 lines)
  - CLI wrapper. Imports from the new core module and re-exports its
    public names for backward compatibility.
  - Adds the repo root to sys.path at import time so
    ``open_alaqs.core.tools.engine_test_import`` is reachable when
    the script is run directly from the repo (not via
    ``python -m``).

* open_alaqs/openalaqsdialog.py
  - New ``OpenAlaqsImportEngineTestEvents`` class (~230 lines).
    Programmatic Qt widgets (no .ui file); layout is a file picker
    row, an "Insert mode" radio group, a "Tolerate warnings"
    checkbox, a read-only monospace text area for the summary, and
    Apply / Close buttons.
  - File choice triggers _run_validation which reads the CSV, runs
    structural + DB-level validation via the core module, and
    displays the formatted summary. The Apply button is enabled
    only when: no errors AND (no warnings OR tolerate-warnings ticked).
  - _on_apply prompts with QMessageBox on the destructive modes:
    * replace-all: strong confirm ("This cannot be undone")
    * replace-for-source: preview of affected source_ids
    matching the CLI's --i-mean-it protection.
  - Uses the same core.tools.engine_test_import.apply_insert so
    numerics match the CLI bit-for-bit.

* open_alaqs/openalaqs.py
  - New ``import_engine_test_events`` action registered next to
    ``taxi_routes`` in the "aircraft related" section of the
    toolbar. Reuses profiles.png as a placeholder icon; a dedicated
    icon is a follow-up.
  - New ``run_import_engine_test_events`` method. Discovers the
    currently-loaded project's DB path via
    ``ProjectDatabase().path`` (same Singleton pattern used by
    OpenAlaqsInventory). Refuses gracefully with a message box if
    no project is loaded.
  - Action enable/disable wired into the existing pattern:
    disabled at plugin load, enabled on project load, disabled on
    project close.

Tests
-----

Existing 39 CLI tests still pass unchanged (they import the
re-exported names from the CLI wrapper).

Verified locally: 56 tests pass across
test_import_engine_test_events.py (39), test_ui_area_sources_toggle.py
(14), test_helicopter_dynamics_by_mode.py (3). No new tests added
for the dialog itself: the interesting logic all lives in the core
module (already fully covered) and the dialog is a thin wrapper
whose only Qt-specific bits are widget hookup. Testing them
meaningfully requires a QGIS session; kept out for now.

Not in this commit
------------------

- No new icon. profiles.png reused as placeholder.
- No .ui file. Widgets built programmatically because the dialog
  is small (5 widgets) and consistent with OpenAlaqsAbout.
- No CSV column for thrust_mode (Phase 4a decision, kept).
